+++ docs/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## [Unreleased]
 
-### Changed
+### Added
 
+- feat: breaks `readState` into `signed` and `unsigned` methods. The `signed` method will take a pre-signed request and re-use it for polling. The `unsigned` method will create a new request and sign it for each poll.
+  - This can be called directly in the `HttpAgent` or through the `Actor` interface, with a new `signed` polling strategy.
 - feat!: removes the unused `defaultAgent` global concept and the `getDefaultAgent` function. The `HttpAgent` constructor is now the only way to create an agent.
 - feat!: removes the `ProxyAgent` class.
 - chore: formatting files and changelog